dragging piece

+++ src/main/scala/example/pages/creator/State.scala
@@ -2,8 +2,9 @@ package example.pages.creator
 
 import com.raquo.laminar.api.L._
 import example.game.Vec2d
+import dev.bluepitaya.laminardragging.Vec2f
 
-case class DraggingPieceState(piece: Models.Piece, position: Vec2d)
+case class DraggingPieceState(piece: Models.Piece, position: Vec2f)
 case class State(
     boardState: BoardContainer.State,
     draggingPieceState: Var[Option[DraggingPieceState]]
